move InstanceEntity defs/impls into entity.rs submodule
